feat(xsnap): high resolution timer: performance.now()
