+++ packages/xsnap/test/test-xsnap.js
@@ -19,7 +19,7 @@ const xsnapOptions = {
   stdout: 'inherit',
 };
 
-function options() {
+export function options() {
   const messages = [];
   async function handleCommand(message) {
     messages.push(decoder.decode(message));
@@ -505,40 +505,3 @@ test('property name space exhaustion: orderly fail-stop', async t => {
     });
   }
 })();
-
-test('meter details', async t => {
-  const opts = options();
-  const vat = xsnap(opts);
-  t.teardown(() => vat.terminate());
-  const result = await vat.evaluate(`
-  let m = new Map();
-  let s1 = new Set();
-  for (ix = 0; ix < 20000; ix++) {
-    m.set(ix, 'garbage');
-    s1.add(ix);
-  }
-  for (ix = 0; ix < 20000; ix++) {
-    m.delete(ix);
-    s1.delete(ix);
-  }
-  `);
-  const {
-    meterUsage: { meterType, ...meters },
-  } = result;
-  t.log(meters);
-  const { entries, fromEntries } = Object;
-  t.deepEqual(
-    {
-      compute: 'number',
-      allocate: 'number',
-      allocateChunksCalls: 'number',
-      allocateSlotsCalls: 'number',
-      garbageCollectionCount: 'number',
-      mapSetAddCount: 'number',
-      mapSetRemoveCount: 'number',
-      maxBucketSize: 'number',
-    },
-    fromEntries(entries(meters).map(([p, v]) => [p, typeof v])),
-  );
-  t.is(meterType, 'xs-meter-6');
-});
